added hide actions to ace actions
